Returning filtering to EmbeddingsWrapper
